Mineral Classification/requirement.txt
numpy==1.21.2
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.1
opencv-python==4.5.3.56
matplotlib==3.4.3
streamlit==0.86.0
pandas==1.3.3
